refactor(base-driver): drop protocol/index.ts import+export duplication, colocate command-name constants with their routes
protocol/index.ts imported every symbol just to re-export it under the same
name; switched to `export {...} from './x.js'` throughout, matching the
convention already used in lib/index.ts.

command-names.ts held five command-name constants disconnected from where
those commands are actually routed. Moved CREATE_SESSION_COMMAND,
DELETE_SESSION_COMMAND, and GET_STATUS_COMMAND into routes/w3c.ts, and
LIST_DRIVER_COMMANDS_COMMAND/LIST_DRIVER_EXTENSIONS_COMMAND into
routes/appium.ts, next to the route definitions that reference them
(replacing the duplicated string literals there too). routes/index.ts
re-exports them from their owning files instead of a standalone module.

No public API changes.
