+++ packages/base-driver/lib/protocol/index.ts
@@ -1,33 +1,15 @@
-import {
-  CREATE_SESSION_COMMAND,
-  DELETE_SESSION_COMMAND,
-  GET_STATUS_COMMAND,
-  LIST_DRIVER_COMMANDS_COMMAND,
-  LIST_DRIVER_EXTENSIONS_COMMAND,
-} from './command-names.js';
-import {errorFromW3CJsonCode, errors, getResponseForW3CError, isErrorType} from './errors.js';
-import {checkParams, makeArgs, validateExecuteMethodParams} from './params.js';
-import {routeConfiguringFunction} from './route-handler.js';
-import {ALL_COMMANDS, METHOD_MAP, NO_SESSION_ID_COMMANDS, routeToCommandName} from './routes/index.js';
-import {isSessionCommand} from './session.js';
-
+export {errorFromW3CJsonCode, errors, getResponseForW3CError, isErrorType} from './errors.js';
+export {checkParams, makeArgs, validateExecuteMethodParams} from './params.js';
+export {routeConfiguringFunction} from './route-handler.js';
 export {
   ALL_COMMANDS,
-  checkParams,
   CREATE_SESSION_COMMAND,
   DELETE_SESSION_COMMAND,
-  errorFromW3CJsonCode,
-  errors,
   GET_STATUS_COMMAND,
-  getResponseForW3CError,
-  isErrorType,
-  isSessionCommand,
   LIST_DRIVER_COMMANDS_COMMAND,
   LIST_DRIVER_EXTENSIONS_COMMAND,
-  makeArgs,
   METHOD_MAP,
   NO_SESSION_ID_COMMANDS,
-  routeConfiguringFunction,
   routeToCommandName,
-  validateExecuteMethodParams,
-};
+} from './routes/index.js';
+export {isSessionCommand} from './session.js';
