+++ packages/base-driver/lib/protocol/routes/w3c.ts
@@ -1,5 +1,9 @@
 import type {Driver, MethodMap} from '@appium/types';
 
+export const CREATE_SESSION_COMMAND = 'createSession';
+export const DELETE_SESSION_COMMAND = 'deleteSession';
+export const GET_STATUS_COMMAND = 'getStatus';
+
 /**
  * Standard W3C WebDriver routes.
  * @see https://www.w3.org/TR/webdriver2/
@@ -8,17 +12,17 @@ export const W3C_ROUTES = {
   // Sessions
   '/session': {
     POST: {
-      command: 'createSession',
+      command: CREATE_SESSION_COMMAND,
       payloadParams: {
         optional: ['capabilities'],
       },
     },
   },
   '/session/:sessionId': {
-    DELETE: {command: 'deleteSession'},
+    DELETE: {command: DELETE_SESSION_COMMAND},
   },
   '/status': {
-    GET: {command: 'getStatus'},
+    GET: {command: GET_STATUS_COMMAND},
   },
 
   // Timeouts
